add backfill capability.
issue https://github.com/raintank/ops/issues/170
adds a new config attribute "backfill" which is the number of seconds
to backfill a metric when it is first produced.  When back-filling,
the first value for the metric is collected, the old data is then
generated so that it is +/- 20% of the collected data.  The variation
of the generated data will be a sine wave.